data: Move selected file change cache
Selected-change state keeps hunk persistence, state snapshots, clear
markers, kind markers, and file-change JSON payloads in one module.

The project lacks a dedicated data owner for selected binary, gitlink,
rename, and text deletion payloads. That spreads file-change serialization
across the same module that owns hunk state and refusal markers.

This commit begins splitting selected-change persistence by moving file-
change JSON readers, loaders, and cache writers into
data.selected_change.file_changes and updating source callers to use that
module directly.

The next commit will update tests and architecture guards to name the new
owner.
