+++ src/git_stage_batch/commands/status.py
@@ -39,12 +39,14 @@ from ..data.progress import format_id_range
 from ..data.selected_change.lifecycle import clear_selected_change_state_files
 from ..data.selected_change.store import (
     SelectedChangeKind,
+    mark_selected_change_cleared_by_stale_batch_selection,
+    read_selected_change_kind,
+)
+from ..data.selected_change.file_changes import (
     load_selected_binary_file,
     load_selected_gitlink_change,
     load_selected_rename_change,
     load_selected_text_deletion_change,
-    mark_selected_change_cleared_by_stale_batch_selection,
-    read_selected_change_kind,
 )
 from ..data.selected_change.snapshots import snapshots_are_stale
 from ..data.session import get_iteration_count
