+++ src/git_stage_batch/data/hunk_tracking.py
@@ -28,6 +28,7 @@ from ..exceptions import NoMoreHunks
 from .auto_advance import resolve_auto_advance
 from . import change_freshness as _change_freshness
 from . import live_diff as _live_diff
+from .selected_change import file_changes as _selected_file_changes
 from .selected_change import store as _selected_store
 from .selected_change import hunk_filtering as _selected_hunk_filtering
 from ..utils.file_io import (
@@ -83,7 +84,7 @@ def fetch_next_change() -> Union[LineLevelChange, BinaryFileChange, GitlinkChang
                     ):
                         continue
 
-                    _selected_store.cache_rename_change(item)
+                    _selected_file_changes.cache_rename_change(item)
                     return item
 
                 if isinstance(item, TextFileDeletionChange):
@@ -97,7 +98,7 @@ def fetch_next_change() -> Union[LineLevelChange, BinaryFileChange, GitlinkChang
                     if is_path_blocked(item.path(), blocked_files):
                         continue
 
-                    _selected_store.cache_text_deletion_change(item)
+                    _selected_file_changes.cache_text_deletion_change(item)
                     return item
 
                 if isinstance(item, GitlinkChange):
@@ -108,7 +109,7 @@ def fetch_next_change() -> Union[LineLevelChange, BinaryFileChange, GitlinkChang
                     if is_path_blocked(item.path(), blocked_files):
                         continue
 
-                    _selected_store.cache_gitlink_change(item)
+                    _selected_file_changes.cache_gitlink_change(item)
                     return item
 
                 # Handle binary files
@@ -122,7 +123,7 @@ def fetch_next_change() -> Union[LineLevelChange, BinaryFileChange, GitlinkChang
                     if is_path_blocked(file_path, blocked_files):
                         continue
 
-                    _selected_store.cache_binary_file_change(item)
+                    _selected_file_changes.cache_binary_file_change(item)
 
                     # Return the BinaryFileChange object directly
                     return item
